Rename prediction modules and refactor lead scoring
