+++ pred_aggregated_amount/catboost_forecast.py
